% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{as.character.R6StringClass}
\alias{as.character.R6StringClass}
\title{as.character}
\usage{
\method{as.character}{R6StringClass}(str)
}
\arguments{
\item{str}{[R6StringClass] string}
}
\value{
[character] The string
}
\description{
Override as.character for R6StringClass
}
\examples{

s   = R6String("Hello")
as.character(s)

}
