% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{as.vector.R6StringClass}
\alias{as.vector.R6StringClass}
\title{as.vector.R6StringClass}
\usage{
as.vector(x,mode)
}
\arguments{
\item{x}{[R6StringClass] string}

\item{mode}{[character]}
}
\value{
[character] base::c(The string)
}
\description{
Override as.vector for R6StringClass
}
\examples{

s   = R6String("Hello")
as.vector(s)

}
