% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{+.R6StringClass}
\alias{+.R6StringClass}
\title{operator +}
\usage{
\method{+}{R6StringClass}(str1, str2)
}
\arguments{
\item{str1}{[R6StringClass] First string}

\item{str2}{[R6StringClass or character] Second string, note that the second can be a character}
}
\value{
[R6StringClass] The string concatenate
}
\description{
Override operator + between two R6StringClass
}
\examples{

s = R6String("Hello")
s = s + " " + "world!" ## "Hello world!"
}
