% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{R6StringClass}
\alias{R6StringClass}
\title{R6StringClass}
\description{
Main class for string in OOP way like python
}
\details{
This class use build-in function from R to provide an interface like python
for string manipulation.
}
\examples{

## Start by defining a string
s = R6String("Hello world!")

print(s) ## print as a character in output
s$replace( " " , "-" ) ## we can replace the space by "-"
s$replace("!","") + " " + "with R6String!" ## Transform to "Hello world with R6String!"
s$lower() ## All in lowercase
s$upper() ## And all in uppercase
R6String("_")$join( "Hello" , "world!" ) ## And we join with sep "_"
s[3:7] ## Access to substring

s = R6String("Hello world! How are you?")
s$title() ## Capitalize for all words

s = R6String("{}_{}")$format("a","b") ## "a_b"

}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{str}}{[character] Underlying string}

\item{\code{size}}{[integer] Length of the string}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{R6StringClass$new()}}
\item \href{#method-print}{\code{R6StringClass$print()}}
\item \href{#method-concatenate}{\code{R6StringClass$concatenate()}}
\item \href{#method-replace}{\code{R6StringClass$replace()}}
\item \href{#method-split}{\code{R6StringClass$split()}}
\item \href{#method-join}{\code{R6StringClass$join()}}
\item \href{#method-upper}{\code{R6StringClass$upper()}}
\item \href{#method-lower}{\code{R6StringClass$lower()}}
\item \href{#method-capitalize}{\code{R6StringClass$capitalize()}}
\item \href{#method-title}{\code{R6StringClass$title()}}
\item \href{#method-format}{\code{R6StringClass$format()}}
\item \href{#method-clone}{\code{R6StringClass$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new R6StringClass object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6StringClass$new(str)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{str}}{[character] String represented by the R6StringClass}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `R6StringClass` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Override the built-in print function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6StringClass$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-concatenate"></a>}}
\if{latex}{\out{\hypertarget{method-concatenate}{}}}
\subsection{Method \code{concatenate()}}{
Concatenate function, used by the operator +
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6StringClass$concatenate(str)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{str}}{[R6StringClass or character] string to concatenate}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-replace"></a>}}
\if{latex}{\out{\hypertarget{method-replace}{}}}
\subsection{Method \code{replace()}}{
Replace pattern `old` by `new`
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6StringClass$replace(old, new)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{old}}{[R6StringClass or character] pattern to replace}

\item{\code{new}}{[R6StringClass or character] new pattern}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-split"></a>}}
\if{latex}{\out{\hypertarget{method-split}{}}}
\subsection{Method \code{split()}}{
Return a list of the words in the string, using sep as the delimiter string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6StringClass$split(sep)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sep}}{[R6StringClass or character] The delimiter}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of R6StringClass
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-join"></a>}}
\if{latex}{\out{\hypertarget{method-join}{}}}
\subsection{Method \code{join()}}{
Concatenate a list of string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6StringClass$join(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{[character or R6String] Any numbers of character or R6String}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-upper"></a>}}
\if{latex}{\out{\hypertarget{method-upper}{}}}
\subsection{Method \code{upper()}}{
Return a copy converted to uppercase
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6StringClass$upper()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
[R6StringClass]
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-lower"></a>}}
\if{latex}{\out{\hypertarget{method-lower}{}}}
\subsection{Method \code{lower()}}{
Return a copy converted to lowercase
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6StringClass$lower()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
[R6StringClass]
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-capitalize"></a>}}
\if{latex}{\out{\hypertarget{method-capitalize}{}}}
\subsection{Method \code{capitalize()}}{
Return a copy capitalized (i.e. the first character is in uppercase, others in lowercase)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6StringClass$capitalize()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
[R6StringClass]
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-title"></a>}}
\if{latex}{\out{\hypertarget{method-title}{}}}
\subsection{Method \code{title()}}{
Transform to title (i.e. all characters after sep are in uppercase, others in lowercase)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6StringClass$title(sep = " ")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sep}}{[R6StringClass or character] separator}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
[R6StringClass]
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format"></a>}}
\if{latex}{\out{\hypertarget{method-format}{}}}
\subsection{Method \code{format()}}{
Format a string by changing all "{}" to values given in l_str
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6StringClass$format(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{[String] String used to format}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
[R6StringClass]
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6StringClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
