% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{[.R6StringClass}
\alias{[.R6StringClass}
\title{operator []}
\usage{
\method{[}{R6StringClass}(str, i)
}
\arguments{
\item{str}{[R6StringClass] string}

\item{i}{[vector of integer] indexes}
}
\value{
[R6StringClass] The substring
}
\description{
Override operator [] two access individual character
}
\examples{

s   = R6String("Hello")
sub = s[2:3] ## "ell"
}
