% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{as.vector.R6StringClass}
\alias{as.vector.R6StringClass}
\title{as.vector}
\usage{
\method{as.vector}{R6StringClass}(str, mode)
}
\arguments{
\item{str}{[R6StringClass] string}
}
\value{
[character] The string
}
\description{
Override as.vector for R6StringClass
}
\examples{

s   = R6String("Hello")
as.vector(s)

}
